introduce deprecation polyfill from another directory
